Create \consistEE wrapper
Provide a wrapper function to install the edition-engraver to an
arbitrary number of contexts with a single call (having to do all
that \layout mangling manually did significantly add complexity to
using the edition-engraver until now).

The function takes a symbol-list with context names and consists
the edition-engraver to all those contexts. Specifying an item that
doesn't map to an existing context triggers a (oll:) warning but doesn't
affect the other contexts.

TODOs:
- the warning about nonexistent context points to the edition-engraver,
  but it should point to the caller.
- It seems that consisting \edition-engraver to different context
  affects the voice count (addressing). This isn't related to the current
  code which just makes that visible. See issue #1